minor edit

+++ pyUtility/schism_file.py
@@ -935,7 +935,7 @@ class schism_grid:
         if not hasattr(self,'bndinfo'): self.compute_bnd()
         if not hasattr(self.bndinfo,'nb'): self.compute_bnd()
         for i in arange(self.bndinfo.nb):
-            fpb=self.bndinfo.ibn[i]; fp=inside_polygon(pxy[sindp],self.x[fpb],self.y[fpb])==1
+            fpb=self.bndinfo.ibn[i].astype('int'); fp=inside_polygon(pxy[sindp],self.x[fpb],self.y[fpb])==1
             sindp=sindp[fp] if self.bndinfo.island[i]==0 else sindp[~fp]
             if len(sindp)==0: break
         sind=zeros(npt).astype('int'); sind[sindp]=1
